requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
redis==5.0.1
python-dotenv==1.0.0
alembic==1.12.1
python-multipart==0.0.6
pytz==2023.3.post1  # 1.3 新增，用于时区处理